backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
itsdangerous>=2.1.0
# Part 2: Gmail API, Drive, scheduler
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
APScheduler>=3.10.0
pytest>=8.0.0
httpx>=0.27.0
